+++ mp_plugin/include/common.h
@@ -1,1 +1,1 @@
-#define _VERSION_ "emulated by CONNECT v0.0.0"
+#define _VERSION_ "emulated by CONNECT v0.2.0"
